backend/chalice/mock_auth_server/requirements.txt
python-jose[cryptography]>=3.1.0
